renamed UserNavbar back to Navbar. Removed searchBar component from itemDetailsPage. Minor styling updates.
