+++ IT-Wizards-UI/src/pages/ItemDetailsPage.jsx
@@ -5,15 +5,9 @@ import SearchResultsList from '../components/search/SearchResultsList';
 
 const ItemDetailsPage = () => {
   const [results, setResults] = useState([]);
-  const [searchTerm, setSearchTerm] = useState('');
 
   return (
     <div>
-      <SearchBar
-        setSearchTerm={setSearchTerm}
-        setResults={setResults}
-        results={results}
-      />
       <SearchResultsList results={results} />
       <ItemDetailDisplay />
     </div>
